docker stuff

+++ docker/install.sh
@@ -2,9 +2,7 @@
 ##############################
 # Init setup
 ##############################
-apt-get install git build-essential vim wget -y # git
-echo "alias l='ls -lah'" >> ~/.bashrc
-echo "alias py='python3'" >> ~/.bashrc
+apt-get install git build-essential vim wget apt-utils systemctl -y
 
 # dir to store tools files
 mkdir ~/tools && cd ~/tools 
@@ -27,9 +25,6 @@ sed -i 's/LIBS := -lz -lpthread/LIBS := -lz -lpthread -lboost_serialization -lhd
 
 make
 
-# add cli command 
-echo "alias ST_BarcodeMap='~/tools/ST_BarcodeMap/./ST_BarcodeMap-0.0.1'" >> ~/.bashrc
-
 ##############################
 # 01.1 FastP: https://github.com/OpenGene/fastp
 ##############################
@@ -37,8 +32,6 @@ cd ~/tools
 
 wget http://opengene.org/fastp/fastp
 chmod a+x ./fastp
-
-echo "alias fastp='~/tools/./fastp'" >> ~/.bashrc
 
 ##############################
 # 01.2 STAR
@@ -49,13 +42,10 @@ rm 2.7.2b.tar.gz
 cd STAR-2.7.2b/source
 make STAR
 
-echo "alias STAR='~/tools/STAR-2.7.2b/source/./STAR'" >> ~/.bashrc
-
 ##############################
 # 02? Bam2Gem
 ##############################
 cd ~/tools
-
 source ~/.bashrc
 
 ##############################
@@ -65,21 +55,22 @@ source ~/.bashrc
 ##############################
 # 03.1 R
 ##############################
-apt-get install gfortran libbz2-dev liblzma-dev libpcre2-dev libcurl4-openssl-dev default-jdk -y
+if false; then
+    apt-get install gfortran libbz2-dev liblzma-dev libpcre2-dev libcurl4-openssl-dev default-jdk -y
 
-wget https://cran.r-project.org/src/base/R-4/R-4.1.2.tar.gz
-tar -xzvf R-4.1.2.tar.gz
-rm R-4.1.2.tar.gz
+    wget https://cran.r-project.org/src/base/R-4/R-4.1.2.tar.gz
+    tar -xzvf R-4.1.2.tar.gz
+    rm R-4.1.2.tar.gz
 
-cd R-4.1.2
-./configure --with-readline=no --with-x=no --enable-R-shlib; make; make install
+    cd R-4.1.2
+    ./configure --with-readline=no --with-x=no --enable-R-shlib; make; make install
+fi
 
 ##############################
 # 03 04 05 07
-#   Seurat, monocle3, SingleR, SciBet, clusterProfiler
+#   Seurat, monocle3, SingleR, SciBet, clusterProfiler dependencies
 ##############################
 apt-get install libssl-dev libxml2-dev libpng-dev -y
-Rscript r_pkgs.R
 
 ##############################
 # 06 CellPhoneDB
@@ -87,16 +78,13 @@ Rscript r_pkgs.R
 apt-get install python3.8 pipenv python3-venv -y
 
 cd ~/tools
-python -m venv cpdb # docs recommend installing in venv
-source cpdb/bin/activate
 pip install cellphonedb
-deactivate
+
 ##############################
-# rStudio Server
+# Aliases
 ##############################
-apt-get install gdebi-core -y
-wget https://download2.rstudio.org/server/bionic/amd64/rstudio-server-2021.09.1-372-amd64.deb
-gdebi rstudio-server-2021.09.1-372-amd64.deb -y
+cat aliases >> ~/.bashrc
+cat aliases >> /home/rstudio/.bashrc
 
 ##############################
 # Finished
